+++ src/main.rs
@@ -1,7 +1,8 @@
 use std::env;
 use std::process;
-use std::fs;
-use std::error::Error;
+
+use grep_rs::Config;
+
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -18,35 +19,8 @@ fn main() {
     println!("Searching for {}", config.query);
     println!("In file {}", config.filename);
 
-    if let Err(e) = run(config) {
+    if let Err(e) = grep_rs::run(config) {
         println!("Application error: {}", e);
         process::exit(1);
     }
-}
-
-
-struct Config {
-    query: String,
-    filename: String,
-}
-
-impl Config {
-    fn new(args: &[String]) -> Result<Config, &'static str> {
-        if args.len() < 3 {
-            return Err("not enough arguments");
-        }
-
-        let query = args[1].clone();
-        let filename = args[2].clone();
-
-        Ok(Config { query, filename })
-    }
-}
-
-fn run(config: Config) -> Result<(), Box<dyn Error>> {
-    let contents = fs::read_to_string(config.filename)?;
-
-    println!("With text:\n{}", contents);
-
-    Ok(())  // 返回 unit 类型 ()
 }
